Update code for better organization of code and conflicts resolution
Updated the codebase to enhance the usability, readability, and the internal working of the various classes. Rearranged the structure of classes for a better organization of the code. Selectors are now present within their respective classes. The addition of system messages in the 'orm' package classes eases the debugging process and provides user with feedback. The 'orm' class 'toString' methods now return more human-readable product, client and warehouse data. Exception handling and error messages improved for better debugging. Fixing conflicts resolution and providing better usability.